Replace transactionId type long with BigInteger (#20)
Issue #24 -  Replace transactionId type long with BigInteger

+++ iron-spi/src/main/java/io/axway/iron/spi/storage/TransactionStore.java
@@ -1,18 +1,19 @@
 package io.axway.iron.spi.storage;
 
 import java.io.*;
+import java.math.BigInteger;
 import java.util.concurrent.*;
 
 public interface TransactionStore {
     OutputStream createTransactionOutput() throws IOException;
 
-    void seekTransactionPoll(long latestProcessedTransactionId);
+    void seekTransactionPoll(BigInteger latestProcessedTransactionId);
 
     TransactionInput pollNextTransaction(long timeout, TimeUnit unit);
 
     interface TransactionInput {
         InputStream getInputStream() throws IOException;
 
-        long getTransactionId();
+        BigInteger getTransactionId();
     }
 }
